# source/dotnet/Auth/Clouded.Auth.Provider/auth-swaggerfile.yml
openapi: 3.0.1
info:
  title: Clouded Auth Provider API
  description: Clouded Auth Provider API
  termsOfService: https://clouded.com/terms
  contact:
    name: Email
    email: support@clouded.com
  version: v1
servers:
  - url: /
paths:
...
  /v1/roles/paginated:
    get:
      tags:
        - Role
      summary: Get all entities paginated
      parameters:
        - name: search
          in: query
          description: ''
          schema:
            type: string
        - name: page
          in: query
          description: ''
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          description: ''
          schema:
            type: integer
            format: int32
            default: 50
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDictionaryPaginatedCloudedOkResult'
  '/v1/roles/by_identity/{identity}':
    get:
      tags:
        - Role
      summary: Get entity by identity
      parameters:
        - name: identity
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: withPermissions
          in: query
          description: ''
          schema:
            type: boolean
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDictionaryCloudedOkResult'
  '/v1/roles/{id}':
    get:
      tags:
        - Role
      summary: Get entity by id
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: withPermissions
          in: query
          description: ''
          schema:
            type: boolean
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDictionaryCloudedOkResult'
    patch:
      tags:
        - Role
      summary: Update entity
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              type: object
              additionalProperties: { }
          text/json:
            schema:
              type: object
              additionalProperties: { }
          application/*+json:
            schema:
              type: object
              additionalProperties: { }
        required: true
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleDictionaryCloudedOkResult'
    delete:
      tags:
        - Role
      summary: Delete entity
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
  '/v1/roles/{id}/permissions/add':
    put:
      tags:
        - Role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolePermissionInput'
          text/json:
            schema:
              $ref: '#/components/schemas/RolePermissionInput'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RolePermissionInput'
        required: true
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
  '/v1/roles/{id}/permissions/remove':
    put:
      tags:
        - Role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolePermissionInput'
          text/json:
            schema:
              $ref: '#/components/schemas/RolePermissionInput'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RolePermissionInput'
        required: true
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudedExceptionResult'
...
components:
  schemas:
    CloudedExceptionResult:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ExceptionData'
          nullable: true
      additionalProperties: false
...
    ExceptionData:
      required:
        - error
        - i18N
        - type
      type: object
      properties:
        i18N:
          type: string
        error:
          type: string
        type:
          type: string
      additionalProperties: false
...
    PageInfo:
      type: object
      properties:
        pageIndex:
          type: integer
          format: int64
        pageSize:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
      additionalProperties: false
...
    RoleDictionaryCloudedOkResult:
      type: object
      properties:
        data:
          type: object
          additionalProperties: { }
          nullable: true
      additionalProperties: false
...
    RoleDictionaryPaginated:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: { }
          description: Paginated data items
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
      additionalProperties: false
    RoleDictionaryPaginatedCloudedOkResult:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RoleDictionaryPaginated'
      additionalProperties: false
    RolePermissionInput:
      type: object
      properties:
        permissionIds:
          type: array
          items: { }
      additionalProperties: false
...
  securitySchemes:
    CloudedKey:
      type: apiKey
      description: Enter the `X-CLOUDED-KEY`
      name: X-CLOUDED-KEY
      in: header
    MachineKey:
      type: apiKey
      description: 'Enter the `X-CLOUDED-MACHINE-KEY` base64 string from Machine api and secret key as following: `[API_KEY]:[SECRET_KEY]`'
      name: X-CLOUDED-MACHINE-KEY
      in: header
    Bearer:
      type: apiKey
      description: 'Enter the Bearer Authorization string as following: `Bearer [TOKEN]`'
      name: Authorization
      in: header
security:
  - CloudedKey: [ ]
  - MachineKey: [ ]
  - Bearer: [ ]